+++ src/validate/validate.ts
@@ -27,7 +27,7 @@ import validateVariableAnchorOffsetCollection from './validate_variable_anchor_o
 import validateSprite from './validate_sprite';
 import ValidationError from '../error/validation_error';
 import validateProjection from './validate_projection';
-import validateProjectionType from './validate_projection-type';
+import validateProjectionType from './validate_projection_type';
 
 const VALIDATORS = {
     '*'() {
