+++ src/validate/validate_projection_type.ts
@@ -2,7 +2,7 @@ import ValidationError from '../error/validation_error';
 import getType from '../util/get_type';
 import v8 from '../reference/v8.json' with {type: 'json'};
 import {ProjectionTypeSpecification, StyleSpecification} from '../types.g';
-import {isPrimitiveProjection, isProjectionTransition, isProjectionType} from '../util/projection';
+import {isPrimitiveProjection, isProjectionType} from '../util/projection';
 
 interface ValidateProjectionOptions {
     sourceName?: string;
